Version 1.0.1 添加了README 修改了日志功能 修复了使用图标无法打包的问题 修复了数据文件会出现空项的问题

+++ build.py
@@ -4,7 +4,7 @@ import os
 import shutil
 import tomllib
 
-VERSION = "1.0.0"
+VERSION = "1.0.1"
 
 config_dict = tomllib.load(open("config.toml", "rb"))
 path_dict = config_dict["Path"]
@@ -19,40 +19,42 @@ os.system(
     fr'cmake -DCMAKE_BUILD_TYPE={"Debug" if is_debug else "Release"} "-DCMAKE_MAKE_PROGRAM={ninja_path}" '
     fr'-DCMAKE_TOOLCHAIN_FILE={vcpkg_path} -DVCPKG_TARGET_TRIPLET=x64-mingw-dynamic -G Ninja -S . -B cmake-build-{build_mode}')
 os.system('cmake --build cmake-build-release --target NuitkaStudio -j 30')
-output_path = f'releases/NuitkaStudio{VERSION}'
-if os.path.exists(output_path):
-    shutil.rmtree(output_path)
-os.mkdir(output_path)
 
-shutil.copyfile("cmake-build-release/NuitkaStudio.exe", f"{output_path}/NuitkaStudio.exe")
-shutil.copyfile("cmake-build-release/libfmt.dll", f"{output_path}/libfmt.dll")
-shutil.copyfile("cmake-build-release/libspdlog.dll", f"{output_path}/libspdlog.dll")
+if not is_debug:
+    output_path = f'releases/NuitkaStudio{VERSION}'
+    if os.path.exists(output_path):
+        shutil.rmtree(output_path)
+    os.makedirs(output_path)
 
-shutil.copyfile("assets/qt.conf", f"{output_path}/qt.conf")
+    shutil.copyfile("cmake-build-release/NuitkaStudio.exe", f"{output_path}/NuitkaStudio.exe")
+    shutil.copyfile("cmake-build-release/libfmt.dll", f"{output_path}/libfmt.dll")
+    shutil.copyfile("cmake-build-release/libspdlog.dll", f"{output_path}/libspdlog.dll")
 
-os.chdir(output_path)
-# remove files
-os.system("windeployqt NuitkaStudio.exe")
-os.remove("Qt5Svg.dll")
-os.remove("libGLESv2.dll")
-os.remove("libEGL.dll")
-os.remove("D3Dcompiler_47.dll")
-os.remove("opengl32sw.dll")
-os.remove("libwinpthread-1.dll")
-os.remove("libgcc_s_seh-1.dll")
-os.remove("libstdc++-6.dll")
+    shutil.copyfile("assets/qt.conf", f"{output_path}/qt.conf")
 
-shutil.rmtree("iconengines")
-shutil.rmtree("imageformats")
-shutil.rmtree("styles")
-shutil.rmtree("translations")
+    os.chdir(output_path)
+    # remove files
+    os.system("windeployqt NuitkaStudio.exe")
+    os.remove("Qt5Svg.dll")
+    os.remove("libGLESv2.dll")
+    os.remove("libEGL.dll")
+    os.remove("D3Dcompiler_47.dll")
+    os.remove("opengl32sw.dll")
+    os.remove("libwinpthread-1.dll")
+    os.remove("libgcc_s_seh-1.dll")
+    os.remove("libstdc++-6.dll")
 
-os.mkdir("qt")
-os.mkdir("qt/plugins")
-shutil.move("platforms", "qt/plugins/platforms")
+    shutil.rmtree("iconengines")
+    shutil.rmtree("imageformats")
+    shutil.rmtree("styles")
+    shutil.rmtree("translations")
 
-shutil.move("Qt5Core.dll", "qt/Qt5Core.dll")
-shutil.move("Qt5Gui.dll", "qt/Qt5Gui.dll")
-shutil.move("Qt5Widgets.dll", "qt/Qt5Widgets.dll")
+    os.mkdir("qt")
+    os.mkdir("qt/plugins")
+    shutil.move("platforms", "qt/plugins/platforms")
+
+    shutil.move("Qt5Core.dll", "qt/Qt5Core.dll")
+    shutil.move("Qt5Gui.dll", "qt/Qt5Gui.dll")
+    shutil.move("Qt5Widgets.dll", "qt/Qt5Widgets.dll")
 
 print(f"Nuitka Studio {VERSION} build successfully!")
